Move ComboboxInput events to hook
