+++ packages/components/src/Form/Inputs/Combobox/index.ts
@@ -32,3 +32,4 @@ export * from './ComboboxContext'
 export * from './ComboboxInput'
 export * from './ComboboxList'
 export * from './ComboboxOption'
+export * from './utils'
