+++ packages/components/src/Form/Inputs/Combobox/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './getComboboxText'
